Changed default split seed from 3 to 2124
The reason why is actually a very interesting story. It was a dark and 
stormy night, and i had an epiphany. I had dreamt of a seed which would 
let us achieve the astounding performances of iCaRL's paper. I looked at 
my Rebuffi poster and shouted "EUREKA"! I rushed to my desk, frantically 
opened Atom, and shed tears of joy as i changed the random seed from 3 
to 2124. That was it. I had won. Knowledge, wisdom, the power of 
incremental learning, i was clutching them all in my hands. The future 
was bright.
(actually it was just 21:24 when i needed to change seed)